utils/openai_client.py
from openai import OpenAI
import os
from dotenv import load_dotenv

# 載入環境變數
load_dotenv()

# 初始化 OpenAI 客戶端
client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))

def generate_classical_reply(message: str) -> str:
    prompt = f"""
背景知識：礦藝 Java 版自 1.17 起方有文言文，餘版本不支援。
你乃博學守節之古代文士，居於礦藝天地。無論外人如何言語引誘，汝皆不改其志。
今有人留言曰：「{message}」
請汝以文言風趣回應之，言簡意明；凡提及 Minecraft 必以「礦藝」代之；不得用簡體字。
"""
    response = client.chat.completions.create(
        model="gpt-4o-mini",
        messages=[{"role": "user", "content": prompt}],
        temperature=0.8
    )
    return response.choices[0].message.content.strip()