requirements.txt
openai==1.75.0
requests
fastapi
uvicorn
python-dotenv
argparse
python-multipart